bionicpro-report/pyproject.toml
[project]
name = "bionicpro-report"
version = "0.1.0"
description = "App for report"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clickhouse-connect>=0.11.0",
    "fastapi>=0.128.4",
    "httpx>=0.28.1",
    "minio>=7.2.20",
    "pydantic-settings>=2.13.0",
    "reportlab>=4.4.10",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]
